+++ pkg/encoding/serialize_yaml.go
@@ -1,7 +1,8 @@
 package encoding
 
 import (
-	"github.com/number571/go-peer/pkg/utils"
+	"errors"
+
 	yaml "gopkg.in/yaml.v2"
 )
 
@@ -12,7 +13,7 @@ func SerializeYAML(pData interface{}) []byte {
 
 func DeserializeYAML(pData []byte, pRes interface{}) error {
 	if err := yaml.Unmarshal(pData, pRes); err != nil {
-		return utils.MergeErrors(ErrDeserialize, err)
+		return errors.Join(ErrDeserialize, err)
 	}
 	return nil
 }