+++ pkg/utils/doc.go
@@ -1,2 +1,0 @@
-// Package utils provides auxiliary functions.
-package utils